fix: Renamed project folder from api_test to booking_api_test fix: Replaced status code asserts with expected_status_code
