+++ booking_api_test/services/auth/auth_api.py
@@ -1,5 +1,5 @@
 import logging
-from api_test.libs.base_api import BaseAPI
+from booking_api_test.libs.base_api import BaseAPI
 
 logger = logging.getLogger(__name__)
 
@@ -17,8 +17,7 @@ class AuthAPI(BaseAPI):
         :return: Return token
         """
         logger.debug(f"Create token: {username=}, {password=}")
-        response = self.post("auth", data={"username": username, "password": password})
-        assert response.status_code == 200, "Access denied"
+        response = self.post("auth", data={"username": username, "password": password}, expected_status_code=200)
         data = response.json()
         logger.debug(f"Auth result: {data}")
         assert "token" in data, "Access denied"
