+++ booking_api_test/services/booking/booking_api.py
@@ -1,6 +1,7 @@
 import requests
-from api_test.libs.base_api import BaseAPI
+from booking_api_test.libs.base_api import BaseAPI
 from .payloads import Payloads
+
 
 class BookingAPI(BaseAPI):
     def __init__(self, *args, **kwargs):
@@ -16,8 +17,8 @@ class BookingAPI(BaseAPI):
         :return: Return API call result as a requests.Response object
         """
         if search and isinstance(search, dict):
-            return self.get("booking", query=search)
-        return self.get("booking")
+            return self.get("booking", query=search, expected_status_code=200)
+        return self.get("booking", expected_status_code=200)
 
     def create_booking(self, data: dict) -> requests.Response:
         """
@@ -26,7 +27,7 @@ class BookingAPI(BaseAPI):
         :param data: booking params (firstname, lastname, etc.)
         :return: API call result as a requests.Response object
         """
-        return self.post("booking", data=data)
+        return self.post("booking", data=data, expected_status_code=200)
 
     def get_booking(self, id: int) -> requests.Response:
         """
@@ -35,10 +36,20 @@ class BookingAPI(BaseAPI):
         :param id: Booking id
         :return: API call result as a requests.Response object
         """
-        return self.get(f"booking/{id}")
+        return self.get(f"booking/{id}", expected_status_code=200)
+
+    def is_booking_exists(self, id: int) -> requests.Response:
+        """
+        Check if booking exists
+
+        :param id: Booking id
+        :return: Return True if booking exists
+        """
+
+        return self.get(f"booking/{id}").status_code == 200
 
     def update_booking(self, id: int, data: dict) -> requests.Response:
-        return self.put(f"booking/{id}", data)
+        return self.put(f"booking/{id}", data, expected_status_code=200)
 
     def delete_booking(self, id: int) -> requests.Response:
-        return self.delete(f"booking/{id}")
+        return self.delete(f"booking/{id}", expected_status_code=201)
